created a root route
when app is build on heroku, its listens to root route to check if app is running or not. So if your app does not return anything on the root route, then heroku considers that as "app crashing" . And it's should be a get method as heroku won't pass any argument or post data.